+++ sites/geohub/src/components/pages/map/Map.svelte
@@ -2,7 +2,6 @@
 	import { page } from '$app/stores';
 	import LayerVisibilitySwitcher from '$components/pages/map/plugins/LayerVisibilitySwitcher.svelte';
 	import MapQueryInfoControl from '$components/pages/map/plugins/MapQueryInfoControl.svelte';
-	import StyleShareControl from '$components/pages/map/plugins/StyleShareControl.svelte';
 	import SplitControl from '$components/util/SplitControl.svelte';
 	import { AdminControlOptions, MapStyles, attribution } from '$lib/config/AppConfig';
 	import { fromLocalStorage, isStyleChanged, storageKeys, toLocalStorage } from '$lib/helper';
@@ -28,10 +27,6 @@
 	import '@maptiler/geocoding-control/style.css';
 	import MaplibreCgazAdminControl from '@undp-data/cgaz-admin-tool';
 	import MaplibreStyleSwitcherControl from '@undp-data/style-switcher';
-	import {
-		MaplibreStaticImageControl,
-		type ControlOptions
-	} from '@undp-data/svelte-geohub-static-image-controls';
 	import { MAPSTORE_CONTEXT_KEY, type MapStore } from '@undp-data/svelte-undp-components';
 	import { SkyControl } from '@watergis/maplibre-gl-sky';
 	import {
@@ -71,7 +66,6 @@
 	let styleSwitcher: MaplibreStyleSwitcherControl;
 
 	export let defaultStyle: string = MapStyles[0].title;
-	let styleUrl = MapStyles[0].uri;
 
 	const layerListStorageKey = storageKeys.layerList($page.url.host);
 	const mapStyleStorageKey = storageKeys.mapStyle($page.url.host);
@@ -99,22 +93,6 @@
 		zoom: 3,
 		hash: true,
 		attributionControl: false
-	};
-
-	let exportOptions: ControlOptions = {
-		width: 300,
-		height: 200,
-		bbox: [-180, -90, 180, 90],
-		latitude: 0,
-		longitude: 0,
-		zoom: 3,
-		bearing: 0,
-		pitch: 0,
-		ratio: 1,
-		defaultApi: 'center',
-		extension: 'png',
-		pageSize: 'A4',
-		orientation: 'landscape'
 	};
 
 	let tourOptions: IntroJsOptions = {
@@ -189,13 +167,23 @@
 				scrollTo: 'off'
 			},
 			{
+				title: 'Exporting map image',
+				intro: `
+            By clicking the export button in layers tab, you can export the current map image with your preferences such as paper size, orientation, file format, etc.
+            `,
+				element: '.tab-layers',
+				position: 'left',
+				step: 5,
+				scrollTo: 'off'
+			},
+			{
 				title: 'Map operations',
 				intro: `
             From this step, we are going to show you main operations on the map.
             `,
 				element: '.map',
 				position: 'floating',
-				step: 5,
+				step: 6,
 				scrollTo: 'off'
 			},
 			{
@@ -205,7 +193,7 @@
             `,
 				element: '.maplibregl-style-switcher-control',
 				position: 'top',
-				step: 6,
+				step: 7,
 				scrollTo: 'off'
 			},
 			{
@@ -215,7 +203,7 @@
             `,
 				element: '.toggle-button',
 				position: 'right',
-				step: 7,
+				step: 8,
 				scrollTo: 'off'
 			},
 			{
@@ -225,29 +213,10 @@
             `,
 				element: '.maplibregl-ctrl-query',
 				position: 'left',
-				step: 8,
-				scrollTo: 'off'
-			},
-			{
-				title: 'Save your work to share with your colleagues',
-				intro: `
-            Once you sign in to your account, this button will be enabled. You can save your current work to share it with your colleagues.
-            `,
-				element: '.maplibregl-ctrl-styleshare',
-				position: 'left',
 				step: 9,
 				scrollTo: 'off'
 			},
-			{
-				title: 'Exporting map image',
-				intro: `
-            You can export the current map image with your preferences such as paper size, orientation, file format, etc.
-            `,
-				element: '.legend-button',
-				position: 'left',
-				step: 10,
-				scrollTo: 'off'
-			},
+
 			{
 				title: 'Disable hillshade layer',
 				intro: `
@@ -255,7 +224,7 @@
             `,
 				element: '.maplibregl-ctrl-hillshade-visibility',
 				position: 'left',
-				step: 11,
+				step: 10,
 				scrollTo: 'off'
 			},
 			{
@@ -265,7 +234,7 @@
             `,
 				element: '.maplibregl-ctrl-geolocate',
 				position: 'left',
-				step: 12,
+				step: 11,
 				scrollTo: 'off'
 			},
 			{
@@ -274,7 +243,7 @@
 					'You have completed map editor tour. Now you can start exploring GeoHub to create a beautiful map. You can always come back to the tour by clicking this button',
 				element: '.tour-control-button',
 				position: 'left',
-				step: 13,
+				step: 12,
 				scrollTo: 'off'
 			}
 		]
@@ -288,8 +257,6 @@
 		const style = $page.data.style;
 		if (style) {
 			// /map/{id} page
-
-			styleUrl = style.links.find((l) => l.rel === 'stylejson').href;
 
 			// if style query param in URL
 			if (`${initiaMapStyleId}` === `${style.id}`) {
@@ -583,17 +550,7 @@
 
 {#if $map}
 	<MapQueryInfoControl bind:map={$map} layerList={layerListStore} position="top-right" />
-	<StyleShareControl bind:map={$map} layerList={layerListStore} position="top-right" />
 	<LayerVisibilitySwitcher bind:map={$map} position="bottom-right" />
-	<MaplibreStaticImageControl
-		bind:map={$map}
-		show={false}
-		style={styleUrl}
-		apiBase={$page.data.staticApiUrl}
-		bind:options={exportOptions}
-		hiddenApiTypes={true}
-		position="top-right"
-	/>
 	<TourControl bind:map={$map} position="top-right" bind:options={tourOptions} />
 {/if}